\docType{data}
\name{dragon_discrete}
\alias{dragon_discrete}

\title{A simulated dataset on the reaction norm of aggressiveness and thermal performance of dragons according to a series of experimental measures at a discretised gradient of temperature}
\format{
A data frame with 1000 rows and 5 variables:
\describe{
\item{Name_Env}{A character string for the "name" of the experimental temperature at which the experiment was performed.}
\item{Env}{The value of the experimental temperature at which the experiment was performed. For convenience, the value was standardised so that the average is 0.}
\item{Individual}{The individual on which the measures were performed at the different temperatures.}
\item{Aggressiveness}{Aggressiveness of the dragon, measured using a very precise, continuous index, when presented with an armored knight to combat.}
\item{Performance}{Performance (sprint speed to rush toward a dummy princess) of each individual dragon evaluated at each of the experimental temperatures.}
}
}
\usage{
dragon_discrete
}
\description{
A simulated dataset on the reaction norm of aggressiveness and thermal performance of dragons according to a series of experimental measures at a discretised gradient of temperature
}
\keyword{datasets}
